paths management is now integrate in span.
send a path to [/span/paths/add] to add this path to the span database.
If the path you send is a file, it will be add as a file, normal !!
If the path you send is a folder, it will add itself as a folder or it
can add all the files inside this folder.
[/span/paths/add/mode] file/folder